+++ tests/Component/Compiler/ConfigurationCompilerTest.php
@@ -7,6 +7,7 @@
 
 namespace GrizzIt\Configuration\Tests\Component\Compiler;
 
+use RuntimeException;
 use PHPUnit\Framework\TestCase;
 use GrizzIt\Vfs\Common\FileSystemInterface;
 use GrizzIt\Vfs\Common\FileSystemDriverInterface;
@@ -27,13 +28,18 @@ class ConfigurationCompilerTest extends TestCase
      * @covers ::__construct
      * @covers ::addLocator
      * @covers ::compile
+     * @covers ::orderPackages
+     *
+     * @runInSeparateProcess
      */
     public function testCompiler(): void
     {
         $driver = $this->createMock(FileSystemDriverInterface::class);
         $subject = new ConfigurationCompiler($driver);
 
+        PackageLocator::registerLocation(__DIR__, 'Foo', ['Bar']);
         PackageLocator::registerLocation(__DIR__);
+        PackageLocator::registerLocation(__DIR__, 'Bar');
 
         $this->assertInstanceOf(ConfigurationCompiler::class, $subject);
 
@@ -43,21 +49,21 @@ class ConfigurationCompilerTest extends TestCase
 
         $fileSystem = $this->createMock(FileSystemInterface::class);
 
-        $driver->expects(static::once())
+        $driver->expects(static::exactly(3))
             ->method('connect')
             ->with(__DIR__)
             ->willReturn($fileSystem);
 
-        $locator->expects(static::once())
+        $locator->expects(static::exactly(3))
             ->method('getLocation')
             ->willReturn('configuration/');
 
-        $fileSystem->expects(static::once())
+        $fileSystem->expects(static::exactly(3))
             ->method('list')
             ->with('configuration')
             ->willReturn(['/foo.json']);
 
-        $fileSystem->expects(static::once())
+        $fileSystem->expects(static::exactly(1))
             ->method('isReadable')
             ->with('configuration/foo.json')
             ->willReturn(true);
@@ -87,5 +93,36 @@ class ConfigurationCompilerTest extends TestCase
             ['foo' => [['foo' => 'bar']]],
             $result->toArray()
         );
+    }
+
+    /**
+     * @return void
+     *
+     * @covers ::__construct
+     * @covers ::addLocator
+     * @covers ::compile
+     * @covers ::orderPackages
+     *
+     * @runInSeparateProcess
+     */
+    public function testCompilerException(): void
+    {
+        $driver = $this->createMock(FileSystemDriverInterface::class);
+        $subject = new ConfigurationCompiler($driver);
+
+        PackageLocator::registerLocation(__DIR__, 'Foo', ['Bar']);
+        PackageLocator::registerLocation(__DIR__, 'Bar', ['Foo']);
+
+        $this->assertInstanceOf(ConfigurationCompiler::class, $subject);
+
+        $locator = $this->createMock(LocatorInterface::class);
+
+        $subject->addLocator($locator);
+
+        $fileSystem = $this->createMock(FileSystemInterface::class);
+
+        $this->expectException(RuntimeException::class);
+
+        $result = $subject->compile();
     }
 }