+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2020-05-25
+### Added
+- The ability to sequence configuration.
+
+### Fixed
+- Prevent loaded multiple of the same configuration.
+
 ## 1.0.1 - 2020-04-29
 ### Fixed
 - Disallowing duplicate values in the registry.
@@ -13,5 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial implementation of the package.
 
 ## Versions
-- [1.0.1 > Unreleased](https://github.com/grizz-it/configuration/compare/1.0.1...HEAD)
+- [1.1.0 > Unreleased](https://github.com/grizz-it/configuration/compare/1.1.0...HEAD)
+- [1.0.1 > 1.1.0](https://github.com/grizz-it/configuration/compare/1.0.1...1.1.0)
 - [1.0.0 > 1.0.1](https://github.com/grizz-it/configuration/compare/1.0.0...1.0.1)